modularize and cleaned up code
